+++ Sources/PeripheryKit/Frontend/Commands/ScanBehavior.swift
@@ -84,13 +84,17 @@ class ScanBehavior {
                         "Periphery is a very precise tool, false positives often turn out to be correct after further investigation."
                 )
             }
+
+            updateChecker.notifyIfAvailable()
+
+            if !filteredDeclarations.isEmpty && configuration.strict {
+                throw PeripheryKitError.foundIssues(count: filteredDeclarations.count)
+            }
         } catch let error as PeripheryKitError {
             return .failure(error)
         } catch {
             return .failure(.underlyingError(error))
         }
-
-        updateChecker.notifyIfAvailable()
 
         if configuration.diagnosisConsole, let graph = result.graph {
             let console = DiagnosisConsole(graph: graph)